ops/monitor/agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyYAML>=6.0
psutil>=5.9.0
